+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
-export const VERSION = "0.6.32";
+export const VERSION = "0.6.33";
 
 export type {
 	CreateComponentArgs,